+++ inits/joomla.js
@@ -5,4 +5,10 @@
  */
 module.exports = {
   name: 'joomla',
+  defaults: {
+    php: '7.4',
+    via: 'apache',
+    xdebug: false,
+    database: 'mysql',
+  },
 };